Split some parts of main into helpers.
